Update to allow connection to single Postgres Database through a single Dataprovider (Vius.Data.DataProvider). Authentication Providers (Membership, Roles, and Capabilities) now all use this dataprovider and pass Postgres style queries.
git-svn-id: file:///home/jeff/Documents/Projects/~repo/Firehall@31 b7b6185a-dcfb-442c-bea6-687e7f8ba4dc

+++ Vius/Web/DbSiteMapNode.cs
@@ -17,7 +17,7 @@ namespace Vius.Web
 	/// </remarks>
 	public class DbSiteMapNode: Vius.Web.SiteMapNode
 	{
-		internal int? ParentKey = null;
+		internal long? ParentKey = null;
 
 		public DbSiteMapNode(DbSiteMapProvider provider, string key)
 			:base(provider, key)
